Pulled out loderunner specific datastep functions.
